block, bfq-sq, bfq-mq: readd missing reset of parent-entity service
The received-service counter needs to be equal to 0 when an entity is
set in service. Unfortunately, commit "block, bfq-sq, bfq-mq: fix
service being wrongly set to zero in case of preemption" mistakenly
removed the resetting of this counter for the parent entities of the
bfq_queue being set in service. This commit fixes this issue by
resetting service for parent entities, directly on the expiration of
the in-service bfq_queue.

Fixes: "block, bfq-sq, bfq-mq: fix service being wrongly set to zero in case of preemption"
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>